manually add changes from #6782
