+++ @xen-orchestra/lite/CHANGELOG.md
@@ -15,7 +15,7 @@
 - Add a star icon near the pool master (PR [#6712](https://github.com/vatesfr/xen-orchestra/pull/6712))
 - Display an error message if the data cannot be fetched (PR [#6525](https://github.com/vatesfr/xen-orchestra/pull/6525))
 - Add "Under Construction" views (PR [#6673](https://github.com/vatesfr/xen-orchestra/pull/6673))
-- Ability to change the state of selected VMs in VMs list (PR [#6782](https://github.com/vatesfr/xen-orchestra/pull/6782))
+- Ability to change the state of selected VMs from the pool's list of VMs (PR [#6782](https://github.com/vatesfr/xen-orchestra/pull/6782))
 
 ## **0.1.0**
